Move and rename GuidelinesColorScreen.kt
